// generated-sources/openapi/model/bookDto.ts
/**
 * Brain Nutrient API
 * API documentation for Brain Nutrient project
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChapterDto } from './chapterDto';


export interface BookDto { 
    id: string;
    name: string;
    whyRead: string;
    chapters: Array<ChapterDto>;
}